fix: resolve all pre-existing clippy warnings and split oversized api.rs
- Extract persistence helpers from api.rs (508 LOC) to new persistence.rs
  (332 LOC remaining, under 500 limit)
- Remove unnecessary Result wrapper from CascadeRetriever::retrieve()
  (always returned Ok, now returns CascadeResult directly)
- Fix clippy::panic + uninlined_format_args in oauth.rs test
- Fix clippy::items_after_statements in episode_handlers.rs
- Fix clippy::cast_precision_loss in batch_operations_demo.rs
- Fix clippy::items_after_statements in oauth_security_tests.rs
- Fix clippy::doc_markdown in episode_update_test.rs

All clippy warnings: 0 | Quality gates: PASS | Tests: 3317 passed
