+++ memory-core/src/retrieval/cascade/tests.rs
@@ -57,7 +57,7 @@ fn test_placeholder_retrieve_without_csm() {
     {
         let mut retriever = CascadeRetriever::default_config();
         retriever.add_episode("ep-1", "Test episode");
-        let result = retriever.retrieve("test query").unwrap();
+        let result = retriever.retrieve("test query");
         // Without CSM feature, returns empty results
         assert!(result.episode_ids.is_empty());
         assert!(result.scores.is_empty());
@@ -120,7 +120,7 @@ mod csm_tests {
         retriever.add_episode("ep-5", "error handling patterns in async code");
 
         // Query with exact keyword match should return 0 API calls
-        let result = retriever.retrieve("authentication JWT token").unwrap();
+        let result = retriever.retrieve("authentication JWT token");
 
         // Should find the matching episode
         assert!(!result.episode_ids.is_empty());
@@ -145,9 +145,7 @@ mod csm_tests {
         retriever.add_episode("ep-5", "fix memory leak in cache implementation");
 
         // Semantic-like query (similar words but not exact match)
-        let result = retriever
-            .retrieve("user authentication and login security")
-            .unwrap();
+        let result = retriever.retrieve("user authentication and login security");
 
         // Should find related episodes
         assert!(!result.episode_ids.is_empty());
@@ -165,7 +163,7 @@ mod csm_tests {
         let retriever = CascadeRetriever::default_config();
 
         // Empty index should indicate API call needed
-        let result = retriever.retrieve("any query").unwrap();
+        let result = retriever.retrieve("any query");
 
         assert!(result.episode_ids.is_empty());
         // Should indicate API fallback needed
@@ -182,11 +180,11 @@ mod csm_tests {
         retriever.add_episode("ep-3", "unique_keyword_gamma optimization");
 
         // Query that matches exactly should hit BM25
-        let result = retriever.retrieve("unique_keyword_alpha").unwrap();
+        let result = retriever.retrieve("unique_keyword_alpha");
         assert!(result.contributing_tiers.contains(&"bm25".to_string()));
 
         // Query with no exact match but similar content should use HDC
-        let result = retriever.retrieve("implement alpha feature").unwrap();
+        let result = retriever.retrieve("implement alpha feature");
         // Either BM25 (if partial match) or HDC (if semantic similarity)
         assert!(!result.episode_ids.is_empty() || result.api_calls > 0);
     }
@@ -212,7 +210,7 @@ mod csm_tests {
         retriever.add_episode("ep-5", "async task spawning performance tips");
 
         // Query that matches multiple aspects
-        let result = retriever.retrieve("Rust async programming").unwrap();
+        let result = retriever.retrieve("Rust async programming");
 
         // Should have results from merging BM25 and HDC
         assert!(!result.episode_ids.is_empty());
@@ -238,7 +236,7 @@ mod csm_tests {
         retriever.add_episode("ep-1", "authentication implementation");
         retriever.add_episode("ep-2", "database connection setup");
 
-        let result = retriever.retrieve("authentication").unwrap();
+        let result = retriever.retrieve("authentication");
 
         // Without merge, should only use single tier
         assert!(result.contributing_tiers.len() <= 1);
@@ -274,7 +272,7 @@ mod csm_tests {
         retriever.add_episode("ep-2", "database pool connection");
         retriever.add_episode("ep-3", "rate limiting API");
 
-        let result = retriever.retrieve("authentication").unwrap();
+        let result = retriever.retrieve("authentication");
 
         // All scores should be in 0.0-1.0 range
         for score in &result.scores {
@@ -295,7 +293,7 @@ mod csm_tests {
             retriever.add_episode(&format!("ep-{i}"), &format!("episode {i} content"));
         }
 
-        let result = retriever.retrieve("episode").unwrap();
+        let result = retriever.retrieve("episode");
 
         // Should not exceed top_k
         assert!(result.episode_ids.len() <= 3);
@@ -350,7 +348,7 @@ mod csm_tests {
         retriever.add_episode("ep-5", "refactor error handling patterns");
 
         // Query with abbreviated terms that need expansion
-        let result = retriever.retrieve("fix auth bug").unwrap();
+        let result = retriever.retrieve("fix auth bug");
 
         // Should find results via concept graph expansion ("auth" → authentication domain)
         assert!(!result.episode_ids.is_empty());
