+++ memory-mcp/examples/batch_operations_demo.rs
@@ -329,7 +329,8 @@ async fn demo_complex_workflow() -> anyhow::Result<()> {
         20 + 80 + 80 + 40 + 60 // Sum of all operations
     );
     println!("  Operations: {}", response.results.len());
-    println!("  Speedup: ~{:.1}x", 280.0 / duration.as_millis() as f64);
+    let speedup = 280.0 / (duration.as_secs_f64() * 1000.0);
+    println!("  Speedup: ~{speedup:.1}x");
 
     // Show execution timeline
     println!("\n  Execution Timeline:");
