+++ memory-mcp/src/bin/server_impl/tools/episode_handlers.rs
@@ -7,6 +7,14 @@ pub async fn handle_bulk_episodes(
 ) -> anyhow::Result<Vec<Content>> {
     use uuid::Uuid;
 
+    #[derive(serde::Serialize)]
+    struct BulkEpisodeResult {
+        requested_count: usize,
+        found_count: usize,
+        missing_count: usize,
+        episodes: Vec<serde_json::Value>,
+    }
+
     let args: Value = arguments.unwrap_or(json!({}));
     let client_id = get_client_id(&args);
 
@@ -76,14 +84,6 @@ pub async fn handle_bulk_episodes(
             serde_json::to_value(ep)
                 .map_err(|e| anyhow::anyhow!("Failed to serialize episode: {e}"))?,
         );
-    }
-
-    #[derive(serde::Serialize)]
-    struct BulkEpisodeResult {
-        requested_count: usize,
-        found_count: usize,
-        missing_count: usize,
-        episodes: Vec<serde_json::Value>,
     }
 
     let bulk_result = BulkEpisodeResult {
